src/main/java/com/algorithm/patterns/simuduck/Duck.java
package com.algorithm.patterns.simuduck;

import com.algorithm.patterns.contstants.Constants;

public abstract class Duck {
    FlyBehavior flyBehavior;
    QuackBehavior quackBehavior;

    public Duck() {
    }

    public abstract String display();

    public void setFlyBehavior(FlyBehavior fb) {
        flyBehavior = fb;
    }

    public void setQuackBehavior(QuackBehavior qb) {
        quackBehavior = qb;
    }

    public String performFly() {
        return flyBehavior.fly();
    }

    public String performQuack() {
        return quackBehavior.quack();
    }

    public String swim() {
        System.out.println(Constants.SWIM);
        return Constants.SWIM;
    }
}
